+++ src/main/java/kr/where/backend/group/dto/groupmember/DeleteGroupMemberDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class DeleteGroupMemberDto {
-    private Long memberIntraId;
+    private Integer memberIntraId;
     private Long groupId;
 
 }
