+++ src/main/java/kr/where/backend/search/SearchService.java
@@ -36,7 +36,7 @@ public class SearchService {
      * 입력 받은 값을 trim으로 공백을 없애주고, 대문자 영어가 들어와도 검색 가능하게 toLowerCase 적용
      */
 
-    public List<ResponseSearch> search(final Long intraId, final String keyWord) {
+    public List<ResponseSearch> search(final Integer intraId, final String keyWord) {
         final String word = validateKeyWord(keyWord.trim().toLowerCase());
         final Member member = memberService.findOne(intraId)
                 .orElseThrow(MemberException.NoMemberException::new);
